Restructured jobsystem basically (WIP)

+++ CreativeChaosEngine/CreativeChaosEditor/src/GUIDrawables/DebugElements/ImGuiDebugInfoLabel.h
@@ -5,7 +5,7 @@ struct ImGuiDebugInfoLabel : public DebugInfoLabel
 {
 public:
 	ImGuiDebugInfoLabel(CCE::String labelName,
-		ImFont* pFont, float* pImGuiDrawTime, ImVec2 size = ImVec2(0, 0),
+		ImFont* pFont, double* pImGuiDrawTime, ImVec2 size = ImVec2(0, 0),
 		ImVec4 bg_color = ImVec4(0.5, 0.5, 0.1, 0.2),
 		ImVec4 text_color = ImVec4(1, 1, 0.2, 0.8f))
 		: DebugInfoLabel(labelName, pFont, false, size, bg_color, text_color), 
@@ -28,6 +28,6 @@ public:
 	}
 
 private:
-	float* pImGuiDrawTime = nullptr;
+	double* pImGuiDrawTime = nullptr;
 
 };
